% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_plots.R
\name{allchains.barplot}
\alias{allchains.barplot}
\title{allchains.barplot}
\usage{
allchains.barplot(X, Y)
}
\arguments{
\item{X}{is a Query ~.allchains object}

\item{Y}{is a Universe ~.allchains object}
}
\description{
this script will generate a barplot to compare the proportion of the specific chains contained in a Query ~.allchains object and in a ~.allchains Universe object.
}
\details{
This barplot will show the percentage of the different unique chains contained in a Query and a Universe ~allchains objects. This function require ggplot2 to generate the graphics.
}
\examples{
allchains.barplot(queryExample.allchains,universeExample.allchains)
}
\author{
Geremy Clair
}
