% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_bionom_test.R
\name{allchains.binom}
\alias{allchains.binom}
\title{allchains.binom()}
\usage{
allchains.binom(X, Y)
}
\arguments{
\item{X}{any <Query.allchains> object}

\item{Y}{any <Universe.allchains> object}
}
\value{
This function will return a data frame with the results of the tests
}
\description{
Performs binom.tests  on an <.allchains> object
}
\details{
Peforms a binomial enrichment test on a <Query.allchains> object compared to the same column of a <Universe.allchains> object.

The universe file db.intact is provided if you are lacking an universe file
}
\examples{
allchains.binom(queryExample.allchains, universeExample.allchains)

}
\author{
Geremy Clair
}
