% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_subset_allchains.R
\name{allchains.cat}
\alias{allchains.cat}
\title{allchains.cat()}
\usage{
allchains.cat(X, Y, test = "Fisher", enrich = FALSE, pval = 0.05,
  adjpval = 1, TGcollapse.rm = TRUE)
}
\arguments{
\item{X}{has to be a <Query>.intact object that was created using the lipid.miner function}

\item{Y}{has to be your <Universe>.intact object that was created using the lipid.miner function}

\item{test}{has to be a type of test in the list "Fisher", "Binom", "Hyper", or "EASE"}

\item{enrich}{TRUE or FALSE (FALSE by default) to choose between showing all the results of the test or only the enriched ones based on a pvalue}

\item{pval}{regular pvalue cutoff (0.05 by default)}

\item{adjpval}{BH adjusted pvalue cutoff (1.0 by default)}

\item{TGcollapse.rm}{Boolean to indicate if the collapsed TG need to be removed (default TRUE)}
}
\value{
This function will return a data.frame containing the tested classifiers (rows), the counts and percentage of identifiers/chains falling in this classifier, a binom.test() pvalue, a BH corrected pvalue and a fold change
}
\description{
Function to evaluate the specific chains enriched within each category
}
\examples{
allchains.cat(queryExample.intact,universeExample.intact,test="Hyper",enrich=TRUE, p=0.05)

}
\author{
Geremy Clair
}
