% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_EASE_test.R
\name{allchains.EASE}
\alias{allchains.EASE}
\title{allchains.EASE()}
\usage{
allchains.EASE(X, Y)
}
\arguments{
\item{X}{any <Query.allchains> object}

\item{Y}{any <Universe.allchains> object}
}
\value{
This function will return a data frame with the results of the tests
}
\description{
Performs EASE.tests  on an <.allchains> object
}
\details{
Peforms an EASE enrichment test (modified Fisher's Exact) on a <Query.allchains> object compared to the same column of a <Universe.allchains> object.

The universe file db.intact is provided if you are lacking an universe file
}
\examples{
allchains.EASE(queryExample.allchains, universeExample.allchains)

}
\author{
Geremy Clair
}
