% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Fisher_test.R
\name{allchains.fisher}
\alias{allchains.fisher}
\title{allchains.fisher()}
\usage{
allchains.fisher(X, Y)
}
\arguments{
\item{X}{any <Query.allchains> object}

\item{Y}{any <Universe.allchains> object}
}
\value{
This function will return a data frame with the results of the tests
}
\description{
Performs fisher exact tests on an <.allchains> object
}
\details{
Peforms a Fisher's Exact enrichment test on a <Query.allchains> object compared to the same column of a <Universe.allchains> object.

The universe file db.intact is provided if you are lacking an universe file
}
\examples{
allchains.fisher(queryExample.allchains, universeExample.allchains)

}
\author{
Geremy Clair
}
