% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_hyper_test.R
\name{allchains.hyper}
\alias{allchains.hyper}
\title{allchains.hyper()}
\usage{
allchains.hyper(X, Y)
}
\arguments{
\item{X}{any <Query.allchains> object}

\item{Y}{any <Universe.allchains> object}
}
\value{
This function will return a data frame with the results of the tests
}
\description{
Performs hyper.tests  on an <.allchains> object
}
\details{
Peforms a hypergeometric enrichment test on a <Query.allchains> object compared to the same column of a <Universe.allchains> object.

The universe file db.intact is provided if you are lacking an universe file
}
\examples{
allchains.hyper(queryExample.allchains, universeExample.allchains)

}
\author{
Geremy Clair
}
