% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{cat.colors}
\alias{cat.colors}
\title{cat.colors}
\description{
cat.colors
}
\details{
This database contains the categories presently supported by Lipid Mini-On and the colors that the figures functions are using.

If you have need another category to be included please contact us
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
