% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Fisher_test.R
\name{chain.fisher}
\alias{chain.fisher}
\title{chain.fisher()}
\usage{
chain.fisher(X, Y)
}
\arguments{
\item{X}{any <Query.chain> object}

\item{Y}{any <Universe.chain> object}
}
\value{
This function will return a data frame with the results of the tests
}
\description{
Performs fisher's exact tests  on an <.allchains> object
}
\details{
Peforms a Fisher's Exact enrichment test on a <Query.chain> object compared to the same column of a <Universe.chain> object.

The universe file db.intact is provided if you are lacking an universe file
}
\examples{
chain.fisher(queryExample.chain, universeExample.chain)

}
\author{
Geremy Clair
}
