% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_hyper_test.R
\name{chain.hyper}
\alias{chain.hyper}
\title{chain.hyper()}
\usage{
chain.hyper(X, Y)
}
\arguments{
\item{X}{any <Query.chain> object}

\item{Y}{any <Universe.chain> object}
}
\value{
This function will return a data frame with the results of the tests
}
\description{
Performs hyper.tests  on an <.chain> object
}
\details{
Peforms a hypergeometric enrichment test on a <Query.chain> object compared to the same column of a <Universe.chain> object.

The universe file db.intact is provided if you are lacking an universe file.
}
\examples{
chain.hyper(queryExample.chain, universeExample.chain)

}
\author{
Geremy Clair
}
