% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_plots.R
\name{chain.length.stack}
\alias{chain.length.stack}
\title{chain.length.stack}
\usage{
chain.length.stack(X)
}
\arguments{
\item{X}{is a ~.chain object}
}
\description{
This function generate a stacked barplot indicating the distribution of the chain lengths within 4 chain lengths categories.
}
\details{
This function will place the chains in four categories based on their length: 1-SCFA (Short-Chain Fatty Acid, 5 or fewer carbons), 2-MCFA (Medium-Chain Fatty Acid, 6-12 carbons), 3-LCFA(Long Chain Fatty Acid, 13-21 carbons), and 4-VLCFA(Very Long Chain Fatty Acid, >22 carbons). This function require ggplot2 to generate the graphics.
}
\examples{
chain.length.stack(queryExample.chain)
}
\author{
Geremy Clair
}
