% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_plots.R
\name{chain.unsat.stack}
\alias{chain.unsat.stack}
\title{chain.unsat.stack}
\usage{
chain.unsat.stack(X)
}
\arguments{
\item{X}{is a ~.chain object}
}
\description{
This function generate a stacked bar plot indicating the distribution of the chain based on their unsaturation count within 4 chain categories.
}
\details{
This function will place the chains in four categories based on their number of unsaturation: 1-Saturated, 2-Monounsaturated, 3-Diunsaturated, and 4-Polyunsaturated(i.e. >2 unsaturations).This function require ggplot2 to generate the graphics.
}
\examples{
chain.unsat.stack(queryExample.chain)
}
\author{
Geremy Clair
}
