% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_clean_lipid_list.R
\name{clean.lipid.list}
\alias{clean.lipid.list}
\title{clean.lipid.list()}
\usage{
clean.lipid.list(X)
}
\arguments{
\item{X}{is a character vector containing the lipids names.}
}
\value{
The R object outputed by this function is a character vector
}
\description{
Clean and prepare a character vector prior to input it in the lipid.miner function.
}
\details{
We recommend to use the clean.lipid.list() prior to use the lipid.miner()

to maximize its compatibility.

This function removes missing values, usless space characters, and redundancy.

clean.lipid.list() informs you using warnings on the steps it has achieved.
}
\examples{
clean.lipid.list(QueryExample)
clean.lipid.list(UniverseExample)

}
\author{
Geremy Clair
}
