% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{cleaned.db}
\alias{cleaned.db}
\title{cleaned.db}
\description{
cleaned.db
}
\details{
This dataset is the result of the clean.lipid.list(db).

In this particular case it just transform db in a character vector

Please see the documentation of db for details on the original db
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
