% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{cleaned.queryExample}
\alias{cleaned.queryExample}
\title{cleaned.queryExample}
\description{
cleaned.queryExample
}
\details{
This dataset is the result of the clean.lipid.list(queryExample).

In this particular case it just transform db in a character vector

Please see the documentation of queryExample for details on the original queryExample
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
