% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{cleaned.universeExample}
\alias{cleaned.universeExample}
\title{cleaned.universeExample}
\description{
cleaned.universeExample
}
\details{
This dataset is the result of the clean.lipid.list(universeExample).

In this particular case it just transform db in a character vector

Please see the documentation of universeExample for details on the original universeExample
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
