% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{db.allchains}
\alias{db.allchains}
\title{db.allchains}
\description{
db.allchains
}
\details{
This dataset is the db.chain object resulting from the lipid.miner function on the cleaned.db object as following:

lipid.miner(cleaned.db,name="db")

Please see the documentation of db and cleaned.db for details on the original database and the lipid.miner() function documentation for the explanation on the .allchains object.
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
