% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{db}
\alias{db}
\title{db}
\description{
db
}
\details{
This database contain a list of 35686 individual lipids with the nomeclature that can be read by lipid Mini-On.

If you have an error when running lipid Mini-On please check that your lipid nomenclature matches with what is in the database.

This object was used to create the following objects: cleaned.db, db.intact, db.chain, and db.allchains
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
