% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{exampleEnrichmentTable}
\alias{exampleEnrichmentTable}
\title{exampleEnrichmentTable}
\description{
exampleEnrichmentTable
}
\details{
This Table contains an example of enrichment result.
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
