% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_bionom_test.R
\name{intact.binom}
\alias{intact.binom}
\title{intact.binom()}
\usage{
intact.binom(X, Y)
}
\arguments{
\item{X}{any column of a <Query.intact> object}

\item{Y}{any column of a <Universe.intact> object (has to be the same column as X)}
}
\value{
This function will return a data frame with the results of the tests
}
\description{
Performs binom.tests  on an <.intact> object column .
}
\details{
Peforms a binomial enrichment test on a column of a <Query.intact>  object compared to the same column of a <Universe.intact> object.

The universe file db.intact is provided if you are lacking an universe file
}
\examples{
intact.binom(queryExample.intact$`Main class`, universeExample.intact$`Main class`)

}
\author{
Geremy Clair
}
