% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_plots.R
\name{intact.cat.stack}
\alias{intact.cat.stack}
\title{intact.cat.stack}
\usage{
intact.cat.stack(X)
}
\arguments{
\item{X}{is a ~.intact object}
}
\description{
This script is meant to produce a stacked bar plot depicting the distribution of the lipid of a given ~.intact object within the different lipid categories.
}
\details{
This function will use the Categories included in the ~.intact object to generate a stacked bar plot of the lipid distribution in a given list. This function require ggplot2 to generate the graphics.
}
\examples{
intact.cat.pie(queryExample.intact)
}
\author{
Geremy Clair
}
