% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_EASE_test.R
\name{intact.EASE}
\alias{intact.EASE}
\title{intact.EASE()}
\usage{
intact.EASE(X, Y)
}
\arguments{
\item{X}{any column of a <Query.intact> object}

\item{Y}{any column of a <Universe.intact> object (has to be the same column as X)}
}
\value{
This function will return a data frame with the results of the tests
}
\description{
Performs EASE.tests  on an <.intact> object column .
}
\details{
Peforms an EASE enrichment test (modified Fisher's Exact) on a column of a <Query.intact>  object compared to the same column of a <Universe.intact> object.

The universe file db.intact is provided if you are lacking an universe file
}
\examples{
intact.EASE(queryExample.intact$`Main class`, universeExample.intact$`Main class`)

}
\author{
Geremy Clair
}
