% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_plots.R
\name{intact.sub.stack}
\alias{intact.sub.stack}
\title{intact.sub.stack}
\usage{
intact.sub.stack(X)
}
\arguments{
\item{X}{is a ~.intact object}
}
\description{
This script is meant to produce a stacked barplot depicting the distribution of the lipid of a given ~.intact object within the different lipid sub classes.
}
\details{
This function will use the Sub classes included in the ~.intact object to generate a stacked barplot of the lipid distribution in a given list. This function require ggplot2 to generate the graphics.
}
\examples{
intact.sub.stack(queryExample.intact)
}
\author{
Geremy Clair
}
