% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_lipid_miner.R
\name{lipid.miner}
\alias{lipid.miner}
\title{lipid.miner()}
\usage{
lipid.miner(X, name = "yourname", TGcollapse.rm = TRUE,
  output.list = FALSE)
}
\arguments{
\item{X}{is character vector containing the lipids names}

\item{name}{is character string setting the name of the <name>.intact, <name>.chain and <name>.allchain object created}

\item{TGcollapse.rm}{is a boolean value. If TRUE, it removes the Triglycerides from the <name>.chain and <name>.allchains files when collapsed}

\item{output.list}{is a boolean value, which defaults to FALSE. If FALSE, the function returns objects as described below. If TRUE, the function instead returns a list of these same objects (minus the appended prefix '<name>.'), thus allowing the user to set the output of this function to a variable of their naming.}
}
\value{
This function return 3 objects named :

<name>.intact, <name>.chain and <name>.allchains

<name>.intact contains : category, main class, subclass, and various chain info

<name>.chain contains information regarding the length of the fatty acid chains

<name>.allchain contains the list of the individual fatty acid
}
\description{
This function performs the text minning and extract lipid information from a <Query> or a <Universe> lipid list. It creates three new distinct R objects in the global environment: <name>.intact, <name>.chain and <name>.allchain
}
\details{
We recommend to use the clean.lipid.list() function prior to use lipid.miner().

lipid.miner() mines the lipid names with the following format:

mainclass(chainLenght:doubleBond/chainLenght:doubleBond/etc.)

the mainclasses recognized by lipid.miner() are contained in the main.colors object

Any lipid that don't belong to one of recognized classes

will be considered as 'Uncategorized'.

Contact us if you need the inclusion of another lipid class
}
\examples{
lipid.miner(cleaned.queryExample, name= "QueryExample", TGcollapse.rm=TRUE)
lipid.miner(cleaned.universeExample, name= "UniverseExample", TGcollapse.rm=TRUE)
}
\author{
Geremy Clair
}
