% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_Network_maker.R
\name{lipid_edge_maker}
\alias{lipid_edge_maker}
\title{lipid_edge_maker()}
\usage{
lipid_edge_maker(X)
}
\arguments{
\item{X}{Any lipid character object, preferably cleaned using the clean.lipid.list() function}
}
\value{
This function will return a data frame with a list of edges
}
\description{
Create a list of lipid classification edges from a cleaned lipid list
}
\details{
Create a list of lipid classification edges from a cleaned lipid list
}
\examples{
Query.edge<-lipid_edge_maker(cleaned.queryExample)

}
\author{
Geremy Clair
}
