% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_Network_maker.R
\name{lipid_network_maker}
\alias{lipid_network_maker}
\title{lipid_network_maker()}
\usage{
lipid_network_maker(X, Y, pval = 1, adjpval = 1, name = "network")
}
\arguments{
\item{X}{either a edge data.frame created with the function edge_maker() or Any lipid character object, preferably cleaned using the clean.lipid.list() function}

\item{Y}{any enrichment table created with Lipid Mini-On}

\item{pval}{unadjusted pvalue cutoff}

\item{adjpval}{BH adjusted pvalue cutoff}

\item{name}{This will be the name of the output of the function ("network" by default")}
}
\value{
This function will return 3 object "name".edge (list of edges), "name".nodes_attributes(list of nodes, type of node, color), "name".edge_attribute (list of edges, colors)
}
\description{
Enable the construction of enrichment networks
}
\details{
Create multiple enrichment network objects based on a query object and a Lipid-Mini-On enrichment analysis table result
}
\examples{
lipid_network_maker(cleaned.queryExample,exampleEnrichmentTable,pval=0.05)

}
\author{
Geremy Clair
}
