% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_Network_maker.R
\name{lipid_visnetwork_plot}
\alias{lipid_visnetwork_plot}
\title{lipid_visnetwork_plot()}
\usage{
lipid_visnetwork_plot(X, Y, pval = 1, adjpval = 1)
}
\arguments{
\item{X}{either a edge data.frame created with the function edge_maker() or Any lipid character object, preferably cleaned using the clean.lipid.list() function}

\item{Y}{any enrichment table created with Lipid Mini-On}

\item{pval}{unadjusted pvalue cutoff}

\item{adjpval}{BH adjusted pvalue cutoff}
}
\value{
This function will plot an interactive network
}
\description{
plots an interactive network of the lipids and enrichment terms
}
\details{
plots an interactive network of the lipids and enrichment terms
}
\examples{
lipid_visnetwork_plot(cleaned.queryExample, exampleEnrichmentTable,pval=0.05)

}
\author{
Geremy Clair
}
