% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{main.colors}
\alias{main.colors}
\title{main.colors}
\description{
main.colors
}
\details{
This database contains the main classes of lipids presently supported by Lipid Mini-On and the colors that the figures functions are using.

If you have need another main class to be included please contact us.
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
