% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{queryExample.intact}
\alias{queryExample.intact}
\title{db.intact}
\description{
db.intact
}
\details{
This dataset is the .intact object resulting from the lipid.miner function on the cleaned.queryExample object as following:

lipid.miner(cleaned.queryExample,name="queryExample")

Please see the documentation of db and cleaned.queryExample for details on the original database and the lipid.miner() function documentation for the explanation on the .intact object
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
