% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{QueryExample}
\alias{QueryExample}
\title{QueryExample}
\description{
QueryExample
}
\details{
This is an example of query list that can be used in Lipid Mini-On

This object was used to create the following objects: cleaned.queryExample, queryExample.intact, queryExample.chain, and queryExample.allchains
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
