% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_QueryWithinUniverse.R
\name{QueryWithinUniverse}
\alias{QueryWithinUniverse}
\title{QueryWithinUniverse()}
\usage{
QueryWithinUniverse(X, Y)
}
\arguments{
\item{X}{a character vector containing a lipid list}

\item{Y}{a character vector containing a lipid list}
}
\value{
This function will return an error message if the all the elements of a <Query>character vector are not a subset of an <Universe>character
}
\description{
Verifies if the lipids of the a <Query> character vector are contained in a <Universe> character vector.
}
\details{
Verifies if the lipids of the a <Query> character vector are contained in a <Universe> character vector.
}
\examples{
QueryWithinUniverse(cleaned.queryExample,cleaned.universeExample)

}
\author{
Geremy Clair
}
