% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_run_the_tests.R
\name{run_the_tests}
\alias{run_the_tests}
\title{run_the_tests()}
\usage{
run_the_tests(Query.miner, Universe.miner, test.type, general.select, subset.by,
  subset.select, enrich, pval, adjpval)
}
\arguments{
\item{Query.miner}{list output from lipid.miner() function run on the Query dataset}

\item{Universe.miner}{list output from lipid.miner() function run on the Universe dataset}

\item{test.type}{character string specifying the type of enrichment test. Valid values are: "Fisher", "Binom", "Hyper", or "EASE".}

\item{general.select}{vector of logical values of length ?, where the vector specifies ? in the following order ?}

\item{subset.by}{character string specifying one of "category", "mainclass", or "subclass", which denote the ?}

\item{subset.select}{vector of logical values of length ?, where the vector specifies ? in the following order ?}

\item{enrich}{logical value specifying whether to ?}

\item{pval}{numeric value specifying the pvalue (between 0 and 1) to use as a cutoff for the enrichment}

\item{adjpval}{numeric value specifying a BH adjusted pvalue (between 0 and 1) to use as a cutoff for the enrichment}
}
\description{
Runs the specified enrichment test(s) and returns a table of results
}
\author{
Geremy Clair, Kelly Stratton
}
