% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{sub.colors}
\alias{sub.colors}
\title{sub.colors}
\description{
sub.colors
}
\details{
This database contains the subclasses of lipids presently supported by Lipid Mini-On and the colors that the figures functions are using.

If you have need another subclass to be included please contact us.
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
