% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_subset_lipidclass.R
\name{subsetcat}
\alias{subsetcat}
\title{subsetcat()}
\usage{
subsetcat(X, cat = "Categoryname")
}
\arguments{
\item{X}{object of class *.intact, *.chain, or *.allchains}

\item{cat}{character string specifying the category of a lipid to subset}
}
\value{
An object of the same class as X.
}
\description{
Subset of the lipid chain object based on it's lipid category.
}
\details{
Subset an object based on its lipid category. Supported categories:

"Sphingolipid","Prenol","Glycerolipid","Sterol","FattyAcyls","Glycerophospholipid"
}
\examples{
subsetcat(queryExample.intact,cat="Glycerolipid")

}
\author{
Geremy Clair
}
