% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_subset_lipidclass.R
\name{subsetsubclass}
\alias{subsetsubclass}
\title{subsetsubclass()}
\usage{
subsetsubclass(X, subclass = "subclassname")
}
\arguments{
\item{X}{object of class *.intact, *.chain, or *.allchains}

\item{subclass}{character string specifying the subclass of lipid to subset}
}
\value{
An object of the same class as X.
}
\description{
Subset of the lipid chain object based on it's lipid subclass
}
\details{
Subset an object based on its lipid main class

(text prior the parenthesis in the lipid name).

The subclasses supported are in the sub.colors object.
}
\examples{
subsetsubclass(queryExample.intact,subclass="LPE(")

}
\author{
Geremy Clair
}
