% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{universeExample.chain}
\alias{universeExample.chain}
\title{universeExample.chain}
\description{
universeExample.chain
}
\details{
This dataset is the .chain object resulting from the lipid.miner function on the cleaned.universeExample object as following:

lipid.miner(cleaned.universeExample,name="universeExample")

Please see the documentation of db and cleaned.universeExample for details on the original database and the lipid.miner() function documentation for the explanation on the .chain object
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
