% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_dataset_documentation.R
\name{UniverseExample}
\alias{UniverseExample}
\title{UniverseExample}
\description{
UniverseExample
}
\details{
This is an example of query list that can be used in Lipid Mini-On

This object was used to create the following objects: cleaned.universeExample, universeExample.intact, universeExample.chain, and universeExample.allchains
}
\author{
Geremy Clair \email{geremy.clair@pnnl.gov}
}
\keyword{data}
